+++ lib/assistant/command.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Assistant
+  class Command
+    attr_reader :content
+
+    def initialize(content)
+      @content = content
+    end
+  end
+end