+++ lib/assistant/commands/linux.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative 'linux/clean'
+require_relative 'linux/stress'