+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## [Unreleased]
 
+
+## [0.1.1] - 2022-09-19
+
+- Add `EH::TickChecklist` command
+- Add `Linux::Clean` command
+- Add `Linux::Stress` command
+
 ## [0.1.0] - 2022-09-11
 
 - Initial release